Fix MODEL 1 gravity vector G definition
During tests with disturbance in X-Y plane, quadrotor drift of MODEL 1
was in opposite direction from all other models. Error was found in
definition of G matrix where wrong rotation matrix was used. Since
gravity is defined as -m*g in inertial frame, it needs to be rotated to
body frame using R_B2E', not R_B2E. This was fixed.