Add controller for X and Y trajectory tracking
Previous models tracked reference on Z axis and only stabilized quads
attitude. To be able to follow a path in space, X and Y tracing was
required. As X and Y cannot be controlled directly (charasteristic of
underactuated system), control is derived using attitude control signals.
This results in control laws of 4th order and achieves tracking but not
disturbance rejection.

ODE solver with variable step is added. It works faster but does not work
with discontinuous signals such and 1st order sliding mode.