Fix estimate of z error derivative; incomplete
Up to now derivative of z error was estimated with filter. This may be
possible and better to do as following: estimate derivative of speed and
subtract reference derivative - which is better is under consideration.

Plot for estimate od z derivative compared it to speed. This is true for
STEP reference when derivative of reference is zero but not true in other
cases. This is to be fixed.