Modify sinusoidal function reference
To make pattern more realistic for quad flight reference is modified in
amplitude and frequency. Next step is to add noise to measured Z to test
realistic reference and measurement with filters and estimators.

Declarations for variables to be integrated and ploted have changed,
this will fail for controllers older that sliding mode (TODO).

Declaration of estimator to be used changed inside super-twisting
controller, this makes easier swithing between using and not using
estimators.