Limit thrust force output to values above zero
Up to now controller was allowed to use negative vertical force if needed
to stabilise. As this is not realistic since motors produce only upward
thrust efficiently, control law is changed in a way that controller is
trying to stabilise around gravity cancellation value (hover)
instead 0 (ground).

Step signal definition is modified to start after 1 second to observe
starting value of vertical force.